embedding_module/requirements.txt
faiss-cpu>=1.8.0
numpy>=1.20
pandas>=1.4
requests>=2.25
PyPDF2>=2.0
gitpython>=3.1
atlassian-python-api>=3.39
beautifulsoup4>=4.11
bs4>=0.0.1
python-docx>=0.8.11
textract>=1.6.5
pillow>=9.0.0
pytesseract>=0.3.10
fastapi>=0.110
uvicorn>=0.24
